Making Specs Actually Match Solution

+++ lib/create.sql
@@ -2,5 +2,5 @@ CREATE TABLE students
     (id INTEGER PRIMARY KEY,
     name TEXT,
     grade INTEGER,
-    gpa INTEGER,
+    gpa FLOAT,
     tardies INTEGER);
